Rename the json-schema example helper files
