+++ packages/constraint-engine/src/json-schema/utils/getRuleConfigurationTupleFromJsonSchema.ts
