+++ packages/constraint-engine/src/json-schema/utils/getTargetReferenceConfigurationTupleFromJson.ts
